Another update to the Camera class. I went and relooked at how we went about positioning the camera, and came up with a different solution based that I feel is a bit easier to understand, and to update. The current method has a range that the X, Y, Z, and Points of interest can go. The position of the players in the world are calculated every frame. As they move away from one another, the camera's Z and Y pull back to keep all players in frame. The x pans when the players move towards either side of the map. I've elected to pan it towards the center of the map. While this does force us to complete a solution to the 'When players are behind a wall' issue sooner, I feel this is the only way to keep enough information on screen to let the players know what's going on. Lastly, I created an offset for the point of interest of the camera which also shifts based on how far away they are from the center of the screen on the Z and X axis. This helps keep the camera focused on the game and avoids having too much dead space when the players are in the corners of the map.
There are some major limitations with the current version. The biggest is that I couldn't find a way to keep all players on screen that kept the players are a reasonable size. As such, I simply elected to complete the original goal; keep them all on screen. We will need to debate how to deal with the size issue soon however. Either we'll need to shrink the game world so there's less ground that could separate them, or we need to adjust the ratios of the rooms. Currently they're square. If we were to shrink their z axis slightly, and make them in roughly 16:9, I think we'd also get around the problem. Will be talking to folks about this ASAP. 

-James Jackson